field.hpp and field_impl.hpp fusion.
