+++ hdl/common/include/util/memory_mapping.hpp
@@ -58,7 +58,6 @@ namespace MemoryMapping {
 #include "memory_mapping/virtual_config.hpp"
 
 #include "memory_mapping/config_impl.hpp"
-#include "memory_mapping/field_impl.hpp"
 #include "memory_mapping/virtual_field_impl.hpp"
 #include "memory_mapping/virtual_config_impl.hpp"
