+++ hdl/common/include/util/memory_mapping/field.hpp
@@ -29,21 +29,35 @@ namespace MemoryMapping {
   public:
     //! \brief Constructor
     //! \param reg : The Register which this Field belongs to
-    constexpr Field(const Register<RegType, Group> reg);
+    constexpr Field(const Register<RegType, Group> reg)
+      : REGISTER(reg) {
+    }
     
     //! \brief Assignment operator for an incompatible Config.
     template<typename OtherRegType, typename OtherGroup, OtherRegType OTHER_MASK>
-    const Field& operator=(const Config<OtherRegType, OtherGroup, OTHER_MASK>& cfg) const;
+    const Field& operator=(const Config<OtherRegType, OtherGroup, OTHER_MASK>& cfg) const {
+      static_assert(!sizeof(RegType) && sizeof(RegType), "Incompatible Field and Config");
+      return *this;
+    }
 
     //! \brief Assignment operator for a Config object.
-    const Field& operator=(const Config<RegType, Group, MASK>& cfg) const;
+    const Field& operator=(const Config<RegType, Group, MASK>& cfg) const {
+      REGISTER = (REGISTER & (RegType)(~MASK)) | cfg.VALUE;
+      return *this;
+    }
 
     //! \brief Assignment operator for an incompatible Field.
     template<typename OtherRegType, typename OtherGroup, OtherRegType OTHER_MASK>
-    const Field& operator=(const Field<OtherRegType, OtherGroup, OTHER_MASK>& field) const;
+    const Field& operator=(const Field<OtherRegType, OtherGroup, OTHER_MASK>& field) const {
+      static_assert(!sizeof(RegType) && sizeof(RegType), "Incompatible Fields");
+      return *this;
+    }
 
     //! \brief Assignment operator for an other Field.
-    const Field& operator=(const Field<RegType, Group, MASK>& field) const;
+    const Field& operator=(const Field<RegType, Group, MASK>& field) const {
+      REGISTER = (REGISTER & (RegType)(~MASK)) | (field.REGISTER & MASK);
+      return *this;
+    }
 
     //! \brief Assignment operator for an integer value.
     //! \param val : An integer value (with some restrictions, see below)
@@ -58,7 +72,10 @@ namespace MemoryMapping {
      *         Then, if the Field is assigned 0b101, the Register's value will be : 0bx1x01xxx
      *         (where "x" stands for the previous value of the Register's corresponding Bit)
      */
-    const Field& operator=(const RegType val) const;
+    const Field& operator=(const RegType val) const {
+      (*this) = make_config(*this, val);
+      return *this;
+    }
 
   };
